Merge Rules into Roles.
After some contemplation it occurred to me that only one thing should ever manage a Role. To have more than one thing able to do so would result in infinite loops. This is a terrible thing to have happen so we need to enforce that a Role is only managed by one Rule. This ends up negating the need to rules at all then so I've merged them together.